requirements.txt
thydra-colorlog==1.2.0
rootutils==1.0.7
transformers==4.29.2
onnx==1.14.0
onnxruntime-gpu==1.15.1
onnxsim==0.4.33
opencv-python==4.8.0.76
segmentation-models-pytorch==0.3.3
ultralytics==8.0.157





hydra-core==1.3.2
hydra-colorlog==1.2.0
hydra-optuna-sweeper==1.2.0

wandb
rootutils
pre-commit
rich
pytest
sh